requirements.txt
pyserial~=3.5
setuptools~=65.5.1
PyAutoGUI~=0.9.53
numpy~=1.24.2